requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML dependencies
openai==1.3.7
langchain==0.0.339
langchain-openai==0.0.2

# Frontend dependencies
streamlit==1.28.1
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.3

# Database
sqlite3

# Utilities
requests==2.31.0
pydantic==2.5.0
typing-extensions==4.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0